+++ structured-log/src/types/logger.interface.ts
@@ -41,3 +41,36 @@ export interface ConsoleComponentLog
      */
     expanded?: boolean;
 }
+
+/**
+ * Definition of log object that is sent to server
+ */
+export interface RestLog
+{
+    /**
+     * Timestamp when log occured
+     */
+    timestamp: string;
+
+    /**
+     * Level of log
+     */
+    logLevel: string;
+
+    /**
+     * Message to be logged
+     */
+    message: string;
+}
+
+/**
+ * Client used for logging logs using REST
+ */
+export interface LoggerRestClient
+{
+    /**
+     * Logs message on server using REST
+     * @param logs - Array of logs to be logged
+     */
+    log(logs: RestLog[]): Observable<void>;
+}