+++ structured-log/src/types/tokens.ts
@@ -2,10 +2,11 @@ import {InjectionToken, Provider, ExistingProvider, ClassProvider, FactoryProvid
 import {LOGGER} from "@anglr/common";
 import {Sink} from "structured-log";
 
-import {ConsoleComponentSink} from "./logger.interface";
-import {ConsoleComponentSinkService} from "../services/consoleComponentSink.service";
+import {ConsoleComponentSink, LoggerRestClient} from "./logger.interface";
+import {ConsoleComponentSinkService} from "../services/console/consoleComponentSink.service";
 import {LoggerService} from "../services/logger.service";
-import {ConsoleSinkConfigService} from "../services/consoleSinkConfig.service";
+import {ConsoleSinkConfigService} from "../services/console/consoleSinkConfig.service";
+import {RestSinkService} from '../services/rest/restSink.service';
 
 /**
  * Factory method for `ConsoleComponentSink`
@@ -26,6 +27,11 @@ export const LOGGER_SINKS: InjectionToken<Sink[]> = new InjectionToken<Sink[]>('
 export const CONSOLE_COMPONENT_SINK_SERVICE: InjectionToken<ConsoleComponentSink> = new InjectionToken<ConsoleComponentSink>('CONSOLE_COMPONENT_SINK_SERVICE');
 
 /**
+ * Injection token for obtaining rest client for rest sink
+ */
+export const LOGGER_REST_CLIENT: InjectionToken<LoggerRestClient> = new InjectionToken<LoggerRestClient>('LOGGER_REST_CLIENT');
+
+/**
  * Provider for ConsoleComponentSinkService for logger
  */
 export const CONSOLE_COMPONENT_SINK_SERVICE_PROVIDER: Provider =
@@ -44,6 +50,17 @@ export const CONSOLE_COMPONENT_SINK: Provider =
 {
     provide: LOGGER_SINKS,
     useExisting: CONSOLE_COMPONENT_SINK_SERVICE,
+    multi: true
+};
+
+/**
+ * Provider for RestSinkService for logger
+ */
+export const REST_SINK: Provider =
+<ClassProvider>
+{
+    provide: LOGGER_SINKS,
+    useClass: RestSinkService,
     multi: true
 };